test(playwright): isolate the collab /replace test in its own project
The test documented that it runs before any editor test, but
fullyParallel guarantees no order: an editor test could run first in the
same worker and its websocket teardown writes race the database reset,
deadlocking that stack's collab container. Workers are never shared
across Playwright projects, so a dedicated project always gives the test
a fresh stack.
